feat: Allow for bodyDef and fixtureDefs to be prepared earlier (#2768)
This change introduces a way to create BodyDef and FixtureDefs as early as in constructor - or even earlier. This allows to pass position and other body/fixture related parameters to constructor, and not have them stick around as this.initialThis or this.initialThat.

For example:

class OffroadWheel extends BodyComponent {
  OffroadWheel(
    Vector2 initialPosition,
    double radius,
  ) : super(
    bodyDef: BodyDef(position: initialPosition), type: BodyType.dynamic),
    fixtureDef: FixtureDef(CircleShape()..radius = radius),
  );
}

I added fixtureDef constructor argument for convenience.