feat: AssetsBundle can be customized in Images and AssetsCache. (#2807)
Previously, `Images` and `AssetsCache` would always try to fetch assets
from the `rootBundle`.

With this PR, we can have instances of those classes that uses a custom
`AssetsBundle`. When this gets merged, we could consider refactoring
flame_network_assets to use this new feature and allow Flame users to
use the same API
to load both local and network assets.