fix: Allow null passthrough parent (#2821)
In `CollisionPassthrough`, while searching for an ancestor with
`CollisionCallbacks` mixin, `firstWhere` was used. This was causing it
to fail when no such ancestor was found. But having a passthrough parent
is not necessary, at least rest of the code wasn't designed on that
assumption. So this PR replaced `firstWhere` with `firstWhereOrNull`
allowing the passthrough to be null.

This was discovered in one of the [discord
discussions](https://discord.com/channels/509714518008528896/516639688581316629/1163600349643411467).