fix: Setting world on FlameGame camera setter (#2831)
# Description

Currently setting a new `FlameGame.camera` can cause `camera.world` to
no long match `game.world`, which I think it's safe to assume is never
desirable.

This PR sets the `camera.world` to the game's `world` when setting a new
Camera, this matches what happens when a new World is set on the game
(`world.camera` is set automatically).