feat: Allow flame tiled to skip tilesets when packing into a tile atlas (#2847)
Adds the ability for Flame tiled to skip tilests in the tile atlas packing. This is useful in case of the map have animation tilesets, which will not be used directly on the tiles, but is still referenced in the tmx file, causing waste of computing resources which can lead to perfomance degradation.