feat: Expose useAtlas on Flame Tiled (#2852)
Flame Tiled uses SpriteBatch under the hood to render its tiles. That API has a flag called useAtlas that when false allows rendering to be done through drawImageRects instead of drawAtlas.

There were not way to set that through Flame Tiled though. This PR exposes that as an option for the developer.