fix: Re-work drags and taps to treat some short drags as taps.
Includes a Base Card to make an empty Stock Pile behave as a Card.
Any attempted drag on a Stock Pile card, including the Base Card,
is changed to a tap in onTapCancel() and the drag is not followed.
In other Piles a short drag is either treated as a tap or ignored.
Only the Waste and Tableau Piles allow such taps. All piles except
the Stock Pile allow drags to start on them.