fix: Allow setting `bounds` while `BoundedPositionBehavior`'s target is null (#2926)
The `bounds` setter of `BoundedPositionBehavior` tries to update the
target's position when bounds are updated. But it wasn't checking if the
target is null. This was causing null exceptions while updating bounds
of an unmounted `BoundedPositionBehavior` with null target (as seen in
[this failing test
case](https://github.com/flame-engine/flame/actions/runs/7231182930/job/19704091006#step:5:1291)).
This PR fixes that by checking if the target is null before updating the
position.


Closes #2655