fix: Fix logic inside flame_riverpod persistent frame callback. (#2950)
In #2943 a persistent frame callback was added to
RiverpodAwareGameWidget's state class, in order to handle multiple
requests in rapid succession to rebuild the widget safely, in response
to [this
thread](https://discord.com/channels/509714518008528896/516639688581316629/1186421112217813062)
on the Discord.

<details>
  <summary>Walkthrough of issue with previous block</summary>
  
**Stage 1:**
forceBuild has been called for the first time, `_isForceBuilding` set to
true

**Stage 2:**
forceBuild has been called a second time before the setState has
completed. `_hasQueuedBuild` is set to true.

**Stage 3:**
The persistent callback is hit. As `_hasQueuedBuild` is true,
`_isForceBuilding` is set to true again, `_hasQueuedBuild` is set to
false / consumed, forceBuild is called again.

**Stage 4:** 
Inside forceBuild, `_isForceBuilding` is still true as it was set inside
the persistent frame callback. Repeat Stage 2
</details>


N/A