Add a hitboxFilter argument to raycast()
Adds a new argument callback to all `raycast*` methods that lets the user ignore hitboxes dynamically. The callback is called with every prospective hitbox, and the hitbox is only considered when the callback returns `true`.

This is faster, in the general case, than the current `ignoreHitboxes` aproach. And it lets the developer have dynamic rules about what the rays collide with. (For example, a line-of-fire raycast should take friendlies into account, but could easily ignore other enemies that get in the way.)

Fixes #2966.