fix: Add Template param to RiverpodGameMixin (#2972)
This PR adds a template parameter to the `RiverpodGameMixin` as it is
done with all other mixins on `FlameGame` throughout the codebase. This
allows to use this mixin on `FlameGame`s that use a different `World`
(for example `Forge2DGame` with `Forge2DWorld`).


I have not created an issue as the solution is straigtforward and this
PR implements the solution.