docs: Fix positioning of git commit message exclusion blocks in the PR template (#3016)
Fix positioning of git commit message exclusion blocks in the PR
template. I found it a bit confusing as I think what we want to
encourage is that the migration instructions go into the git message,
but rn there is just a "Replace or remove this text." lost after the
related issues. I don't think we want to include related issues? So it
seems just out of order.

---------

Co-authored-by: Lukas Klingsbo <me@lukas.fyi>