fix: Fix cascading and fallback propagation of text styles (#3129)
Fix cascading and fallback propagation of text styles.

While using text nodes and flame_markdown, I noticed that basic
propagating of styles was not working.

For example, if you create a standard document, it will have some
margins by default. Even when setting it to zero, it was still
overwritten by the default style.

I noticed that there was no consistency in the methods `copyWith` and
`merge` of which parameter/receiver should be the master and which
should be the fallback. Both the implementations and the callers would
use it either way indiscriminately.

This consolidates the definition and adds comments to the methods
enforcing the priority order. Also fixes the wrong precedence on
`copyWith` methods.

In order to fully test this, I had to expose a few internal details of
text elements; I think this is fine since Dart supports the
`@visibleForTesting` annotation, and it makes the tests cleaner.