fix: Make AdvancedButtonComponent call onPressed in onTapUp (#3150)
`AdvancedButtonComponent` doesn't show the `downSkin` if the action of `onPressed` changes the router to another page. That is because the router is changed before the `downSkin` is shown as the `onPressed` action is called in `onTapDown` instead of `onTapUp`.

In this PR, calling `onPressed` action is moved to `onTapUp` method instead of `onTapDown`. That makes the `AdvancedButtonComponent` more natural and allows users to skip the action if they move the mouse pointer or their finger outside the button box.