feat: Add `onReleased` action to `AdvancedButtonComponent` which will be called within `onTapUp` (#3152)
Add `onReleased` action which will be called within `onTapUp`. That
fixed some scenarios which require the pressed action to be called after
the user taps up the button.

For example, `AdvancedButtonComponent` doesn't show the `downSkin` if
the action of `onPressed` changes the router to another page. That is
because the router is changed before the `downSkin` is shown as the
`onPressed` action is called in `onTapDown`.