fix: Invoke `setToStart` on child effect controller of wrapping effect controllers (#3168)
`DelayedEffectController` was not calling `setToStart` on its child,
causing it to not behave as expected when used with
`InfiniteEffectController`. This was reported by a discord member [in
this
message](https://discord.com/channels/509714518008528896/516639688581316629/1242258377766080666).

This PR introduces a `HasSingleChildEffectController` mixin which can be
added to any effect controller that wraps a single child effect
controller. This mixin makes sure that the `setToStart`, `setToEnd` and
`onMount` methods always get invoked on the child controllers.