feat: Expose vector classes on core file [flame_3d] (#3211)
Expose vector classes on extensions file.

It is very common that we want to work with vector and matrices even
though I don't want to import anything else specific from Flame. This PR
allows we to use the extensions file as a base import similar to how it
is on normal flame.