test: Always clean up after failed test (#3307)
When running several `testWithGame` (or `testWithFlameGame`) tests after
one another, if one test failed with a thrown error (e.g. failed
`expect()`), the following ones would fail as well with failed assertion
`Component.staticGameInstance == null`. This is because
`game.onRemove()` didn't get to run after the thrown error.

This polluted the output and made it hard to understand which test
actually failed and which didn't. It was also confusing to new users.