docs: Update step_1.md to include imports for Canvas (#3351)
I have added the import required for the Canvas component. This tripped
me up originally as I was following this example and had to go find out
what import to use for the line `void render(Canvas canvas) {`

I did not make an issue since this is an extremely small issue and is
regarding documentation and not any real code.