docs: Remove counterclockwise order requirement for PolygonComponent vertices (#3352)
The documentation used to require that lists of vertices in
PolygonComponent (and its subclass PolygonHitbox) should be in
counterclockwise order. This is no more necessary because the
implementation is robust:
https://github.com/flame-engine/flame/blob/main/packages/flame/lib/src/geometry/polygon_component.dart#L122-L124

Moreover, the wording was a bit confusing (the example looks clockwise).

The paragraphs mentioning the vertex order were deleted altogether.