chore: Update flame_3d CHANGELOG.md to match main + changes [flame_3d] (#3380)
Update flame_3d CHANGELOG.md to match main + changes.

This is the second step in merging flame_3d into main. Before we do a
final rebase of main back, the last "conflict" is the changelog file (as
can be seen [on the temporary PR
diff](https://github.com/flame-engine/flame/pull/3377)).

In order to solve this conflict, this does the following steps:
* Collect the changes from flame_3d compared to the current base branch
version of main (i.e.: everything that was added to the changelog; this
turns out to be two entries for the same day, `2024-09-16`).
* Replace the CHANGELOG.md file with the latest from current main
* Add back the missing entries, but I have doctored the days to be today
(`2024-12-10`), in order to not "alter the past". I am happy with
keeping the original date, though. No strong preference either way.

Once this is merged, the diff PR should show no conflict, and we should
be able to finally rebase it with current main to get rid of all the
extraneous files. Finally, we can create a new merge PR and the diff
should be exactly what we want.