fix: Unmultiply before brightening (#3407)
The image data read back is in a pre-multiplied state. Convert it back
to straight alpha before brightening and then premultiply.

This just makes the colors go to white though:

![image](https://github.com/user-attachments/assets/1b38f1ed-85ec-4ac5-abc4-4ba7c5392173)

You could alternatively use the HSV color space and increase the Value
(brightness); this would be more subtle and keep the colors correct.


![image](https://github.com/user-attachments/assets/a87780ab-6698-46cc-aad6-ca37baad229c)