test: Enhance tests for flame_markdown to ensure nested styles are propagated (#3439)
Enhance tests for flame_markdown to ensure nested styles are propagated.

Specifically adds a second section to each test to assert the element
tree (after having asserted the node tree).

And then uses that to add a nested inline style test making sure the
"CSS" is propagated and merged correctly.