feat: Add multiFactory to SpawnComponent (#3440)
add multiFactory to SpawnComponent to be able to create a list of
components for each spawn tick. For backwards compatibility the single
component factory is still supported and wrapped internally.


This PR enhances the SpawnComponent to accept a multiFactory which adds
a list of components per tick instead of just one.

---------

Co-authored-by: Patrick Cornelißen <patrick.cornelissen@ars.de>
Co-authored-by: Luan Nico <luanpotter27@gmail.com>
Co-authored-by: Lukas Klingsbo <lukas.klingsbo@gmail.com>