fix: Only expose `ReadOnlyOrderedset` from `component.children` (#3606)
This PR updates to ordered_set 8.0.0 and makes sure that `children` are
of the `ReadOnlyOrderedSet` type, so that `children.add/remove` isn't
directly exposed to the user, since they skip the component lifecycle.

Since we previously had the `ComponentSet` to hide `add/remove` we now
had to introduce a new internal getter in the component to make sure
that the set only is initialized when it is needed. This was a previous
optimization to keep components without children more lightweight.

It also introduces the static `Component.strictQueryMode` variable so
that the user can set the `ComponentSetFactory` to only create
components sets with strict mode turned on (this functionality also
previously lived in the `ComponentSet` class).